requirements.txt
# PyTorch and related libraries
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Transformers and Hugging Face
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Vector quantization
vector-quantize-pytorch>=1.8.0

# CLAP model
laion-clap>=1.1.0

# Diffusion models
diffusers>=0.18.0

# Music source separation (Demucs)
demucs>=4.0.0

# Audio analysis
essentia>=2.1b6.dev1084

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Evaluation metrics
mir-eval>=0.7
museval>=0.4.0

# Development tools
jupyter>=1.0.0
ipython>=8.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional dependencies for enhanced functionality
# Uncomment if needed

# Fast tokenizers
# tokenizers>=0.13.0

# GPU acceleration
# nvidia-ml-py3>=7.352.0

# Advanced audio processing
# pyworld>=0.3.0
# crepe>=0.0.12

# Speech recognition (for lyrics extraction)
# whisper-openai>=20230314
# faster-whisper>=0.6.0

# Music information retrieval
# madmom>=0.16.1
# aubio>=0.4.9

# Alternative audio libraries
# pedalboard>=0.7.0
# pydub>=0.25.0

# Memory optimization
# psutil>=5.9.0

# Distributed training
# deepspeed>=0.9.0